// src/app/models/proposalModel.ts
export class proposalModel{
    userId:String = "";
    //typeDocument:String = "";
    entity:String = "";
    proposal: boolean = true;
    sendBy:String ="";
    sendByEmail:String ="";
    reference: String = "";
    typeProposal: String = "";
    clientQuery: boolean = false;

    fileBase64:String = "";
    fileName: String = "";
    fileExtention:String = "";
    description: String = "";
    amount: number = 0;
    bank:String = "";

  
}